requirements.txt
nni
torch
numpy
matplotlib
biopython
pandas
torchvision
torchaudio
pytorch_lightning
tqdm
scikit-learn